Working movement, shooting and looking around.
Had to do away with the gesture recognizers, they didn't seem to do what was needed.  Likely have to track which touch came from which view, as movement turns into looking when you move beyond the boarders of the view.  Also would be nice to start shooting and look around to implement indiscriminant firing.