Initial working RootViewController implementation
Serious bug, in that GomeViewController had a @synthesize view.  This
prevents many of the methods of UIViewController from working properly
in subclasses.  Fixed, and added new layer to view hierarchy.

Seems that the main menu displays properly, but the game engine
doesn't (it's rotated 90 degrees).  Should be simple to fix from here.